refactor: reestruturei o código para uma arquitetura mais organizada v2
