+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Decode/Chassis.java
@@ -1,6 +1,7 @@
-package org.firstinspires.ftc.teamcode.IntoTheDeep;
+package org.firstinspires.ftc.teamcode.Decode;
 
 import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
@@ -12,6 +13,7 @@ import org.firstinspires.ftc.robotcore.external.navigation.AxesOrder;
 import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
 import org.firstinspires.ftc.robotcore.external.navigation.Orientation;
 
+@Disabled
 @TeleOp
 public class Chassis extends OpMode {
     DcMotor motorFL;
