+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Util/FODChassis.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.IntoTheDeep;
+package org.firstinspires.ftc.teamcode.Util;
 
 import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
@@ -6,12 +6,14 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.IMU;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.AxesOrder;
 import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
 import org.firstinspires.ftc.robotcore.external.navigation.Orientation;
 
+//@Disabled
 @TeleOp
 public class Chassis extends OpMode {
     DcMotor motorFL;
@@ -52,13 +54,11 @@ public class Chassis extends OpMode {
         motorFR = hardwareMap.get(DcMotor.class, "FR");
         motorBL = hardwareMap.get(DcMotor.class, "BL");
         motorBR = hardwareMap.get(DcMotor.class, "BR");
-        //motorV = hardwareMap.get(DcMotor.class, "V");
 
         motorFL.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
         motorFR.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
         motorBL.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
         motorBR.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-        //motorV.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
 
         motorFL.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
         motorFR.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
@@ -67,12 +67,6 @@ public class Chassis extends OpMode {
 
         motorFL.setDirection(DcMotorSimple.Direction.REVERSE);
         motorBL.setDirection(DcMotorSimple.Direction.REVERSE);
-        //motorV.setDirection(DcMotorSimple.Direction.REVERSE);
-
-        //motorV.setTargetPosition(0);
-        //motorV.setPower(1);
-
-        //motorV.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
         telemetry.addData("Hardware: ", "Initialized");
     }
@@ -84,12 +78,6 @@ public class Chassis extends OpMode {
     public void loop(){
         telemetry.addData("Hardware: ", "Running");
 
-        /* if (gamepad1.y) {
-            motorV.setTargetPosition(4200);
-        } else if (gamepad1.a) {
-            motorV.setTargetPosition(0);
-        } */
-
         double drive = -gamepad1.left_stick_y;  // frente e atrás
         double turn = -gamepad1.right_stick_x;  // gira
         double strafe = gamepad1.left_stick_x; // direita e esquerda
@@ -100,7 +88,6 @@ public class Chassis extends OpMode {
         if(gamepad1.left_bumper) imu.resetYaw();
 
         telemetry.addLine("Angulo do robô: "+ imu.getRobotYawPitchRollAngles().getYaw());
-        //telemetry.addLine("Viper: "+motorV.getCurrentPosition());
 
         double heading = imu.getRobotYawPitchRollAngles().getYaw(AngleUnit.RADIANS);
 
@@ -111,6 +98,7 @@ public class Chassis extends OpMode {
         motorFR.setPower(((adjustedDrive - adjustedStrafe - turn) / max) * drivePower);
         motorBL.setPower(((adjustedDrive - adjustedStrafe + turn) / max) * drivePower);
         motorBR.setPower(((adjustedDrive + adjustedStrafe - turn) / max) * drivePower);
+
         telemetry.update();
     }
 }
